CommittedMemoryProvider cleanups.
